+++ src/main/java/org/teamsai/saibackend/domain/matching/service/AutoMatchingService.java
@@ -4,13 +4,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.teamsai.saibackend.domain.matching.exception.MatchingErrorCode;
-import org.teamsai.saibackend.domain.matching.model.AutoMatchingExecutionResult;
-import org.teamsai.saibackend.domain.matching.model.AutoMatchingResult;
-import org.teamsai.saibackend.domain.matching.model.AutoMatchingTransactionResult;
-import org.teamsai.saibackend.domain.matching.model.EvaluatedMatchingCandidate;
-import org.teamsai.saibackend.domain.matching.model.MatchingCandidate;
-import org.teamsai.saibackend.domain.matching.model.MatchingTransaction;
-import org.teamsai.saibackend.domain.matching.policy.AutoMatchingJudge;
+import org.teamsai.saibackend.domain.matching.service.AutoMatchingExecutionResult;
+import org.teamsai.saibackend.domain.matching.service.AutoMatchingResult;
+import org.teamsai.saibackend.domain.matching.service.AutoMatchingTransactionResult;
+import org.teamsai.saibackend.domain.matching.service.EvaluatedMatchingCandidate;
+import org.teamsai.saibackend.domain.matching.service.MatchingCandidate;
+import org.teamsai.saibackend.domain.matching.service.MatchingTransaction;
+import org.teamsai.saibackend.domain.matching.service.AutoMatchingJudge;
 import org.teamsai.saibackend.domain.matching.type.AutoMatchingProcessStatus;
 import org.teamsai.saibackend.domain.matching.type.MatchingTargetType;
 import org.teamsai.saibackend.domain.payment.exception.PaymentErrorCode;
