+++ src/main/java/org/teamsai/saibackend/domain/matching/service/BankMatchingTransactionService.java
@@ -4,10 +4,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.teamsai.saibackend.domain.matching.exception.MatchingErrorCode;
 import org.teamsai.saibackend.domain.matching.dto.BankTransactionMatchCandidateDTO;
-import org.teamsai.saibackend.domain.matching.model.AutoMatchingExecutionResult;
-import org.teamsai.saibackend.domain.matching.model.AutoMatchingTransactionResult;
-import org.teamsai.saibackend.domain.matching.model.MatchingCandidate;
-import org.teamsai.saibackend.domain.matching.model.MatchingTransaction;
+import org.teamsai.saibackend.domain.matching.service.AutoMatchingExecutionResult;
+import org.teamsai.saibackend.domain.matching.service.AutoMatchingTransactionResult;
+import org.teamsai.saibackend.domain.matching.service.MatchingCandidate;
+import org.teamsai.saibackend.domain.matching.service.MatchingTransaction;
 import org.teamsai.saibackend.domain.matching.type.AutoMatchingProcessStatus;
 import org.teamsai.saibackend.domain.matching.type.AutoMatchingTransactionType;
 import org.teamsai.saibackend.domain.matching.type.MatchingTargetType;
