+++ src/main/java/org/teamsai/saibackend/domain/payment/mapper/PaymentObligationMapper.java
@@ -2,7 +2,7 @@ package org.teamsai.saibackend.domain.payment.mapper;
 
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
-import org.teamsai.saibackend.domain.matching.model.MatchingCandidate;
+import org.teamsai.saibackend.domain.matching.service.MatchingCandidate;
 import org.teamsai.saibackend.domain.matching.type.MatchingTargetType;
 import org.teamsai.saibackend.domain.payment.dto.PaymentObligationDTO;
 import org.teamsai.saibackend.domain.payment.type.PaymentStatus;
