+++ src/test/java/org/teamsai/saibackend/domain/matching/BankMatchingServiceTest.java
@@ -7,8 +7,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.teamsai.saibackend.domain.matching.exception.MatchingErrorCode;
-import org.teamsai.saibackend.domain.matching.model.AutoMatchingExecutionResult;
-import org.teamsai.saibackend.domain.matching.model.AutoMatchingTransactionResult;
+import org.teamsai.saibackend.domain.matching.service.AutoMatchingExecutionResult;
+import org.teamsai.saibackend.domain.matching.service.AutoMatchingTransactionResult;
 import org.teamsai.saibackend.domain.matching.service.BankMatchingService;
 import org.teamsai.saibackend.domain.matching.service.BankMatchingTransactionService;
 import org.teamsai.saibackend.domain.matching.type.AutoMatchingProcessStatus;
