+++ src/test/java/org/teamsai/saibackend/domain/matching/BankMatchingTransactionServiceTest.java
@@ -8,10 +8,10 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.teamsai.saibackend.domain.matching.exception.MatchingErrorCode;
 import org.teamsai.saibackend.domain.matching.dto.BankTransactionMatchCandidateDTO;
-import org.teamsai.saibackend.domain.matching.model.AutoMatchingExecutionResult;
-import org.teamsai.saibackend.domain.matching.model.AutoMatchingTransactionResult;
-import org.teamsai.saibackend.domain.matching.model.MatchingCandidate;
-import org.teamsai.saibackend.domain.matching.model.MatchingTransaction;
+import org.teamsai.saibackend.domain.matching.service.AutoMatchingExecutionResult;
+import org.teamsai.saibackend.domain.matching.service.AutoMatchingTransactionResult;
+import org.teamsai.saibackend.domain.matching.service.MatchingCandidate;
+import org.teamsai.saibackend.domain.matching.service.MatchingTransaction;
 import org.teamsai.saibackend.domain.matching.service.AutoMatchingService;
 import org.teamsai.saibackend.domain.matching.service.BankMatchingTransactionService;
 import org.teamsai.saibackend.domain.matching.service.BankTransactionMatchCandidateService;
